Completely revamped how the program paints functions
Functions are now added to an editable list. Graph color can also be set
as well as axis color. The way the program paints the canvas is now more
advanced, and all active functions are saved in a list which gets
repainted each time a new function is added or edited. Color can not yet
be edited once the function has been added to the list.